Fix Provider API types.

+++ provider/src/main/java/com/seagate/alto/provider/lyve/response/Location.java
@@ -13,10 +13,10 @@ public class Location {
 
     @SerializedName("latitude")
     @Expose
-    public Integer latitude;
+    public Double latitude;
     @SerializedName("longitude")
     @Expose
-    public Integer longitude;
+    public Double longitude;
 
     @Override
     public String toString() {
